Record safety checks

+++ spynnaker/pyNN/models/populations/population_view.py
@@ -499,7 +499,7 @@ class PopulationView(PopulationBase):
 
     @overrides(PopulationBase.record)
     def record(self, variables: Names, to_file: IoDest = None,
-               sampling_interval: Optional[int] = None) -> None:
+               sampling_interval: Optional[float] = None) -> None:
         self.__recorder.record(
             variables, to_file, sampling_interval, self.__indexes)
